+++ app/src/main/java/yokai/core/di/PreferenceModule.kt
@@ -13,6 +13,7 @@ import org.koin.dsl.module
 import yokai.domain.backup.BackupPreferences
 import yokai.domain.base.BasePreferences
 import yokai.domain.download.DownloadPreferences
+import yokai.domain.library.LibraryPreferences
 import yokai.domain.recents.RecentsPreferences
 import yokai.domain.source.SourcePreferences
 import yokai.domain.storage.StoragePreferences
@@ -46,6 +47,8 @@ fun preferenceModule(application: Application) = module {
     single { SecurityPreferences(get()) }
 
     single { BackupPreferences(get()) }
+
+    single { LibraryPreferences(get()) }
 
     single {
         PreferencesHelper(