+++ app/src/main/java/yokai/domain/library/LibraryPreferences.kt
@@ -1,0 +1,7 @@
+package yokai.domain.library
+
+import eu.kanade.tachiyomi.core.preference.PreferenceStore
+
+class LibraryPreferences(private val preferenceStore: PreferenceStore) {
+    fun randomSortSeed() = preferenceStore.getInt("library_random_sort_seed", 0)
+}